create driver_example and move code from current_monitoring to it
